kconfig/satconfig: add satconfig program
This commit adds support for a new target, "make satconfig".

satconfig is a new program that takes a partial assignment of CONFIG_*
variables and tries to find an assignment that will satisfy all the
constraints of the kernel's Kconfig files.

The partial assignment is read from a file called ".satconfig" in the
top-level directory. For example, if your .satconfig contains only the
following:

  CONFIG_EXT4_FS=m

and you run "make satconfig", this will output a .config file where ext4
is enabled (as a module) and all its dependencies are enabled as well.
The remaining config options will be disabled by default, thus enabling
the user to specify only those options they truly care about while not
building anything else.

Thanks & shoutouts to:

* Bart Massey for supervising my GSoC'10 project
* Sam Ravnborg for explaining some of the trickier parts of kconfig
* Armin Biere for PicoSAT support and assistance

Cc: Bart Massey <bart.massey@gmail.com>
Cc: Sam Ravnborg <sam@ravnborg.org>
Cc: Masahiro Yamada <masahiroy@kernel.org>
Cc: Armin Biere <biere@cs.uni-freiburg.de>
Cc: Andi Kleen <ak@linux.intel.com>
Cc: Jan Beulich <JBeulich@suse.com>
Cc: James Bottomley <jejb@linux.ibm.com>
Cc: Felipe Contreras <felipe.contreras@gmail.com>
Link: https://lore.kernel.org/lkml/AANLkTil73ArhX2w6b_v1bmWjP1FA5zUTfPzK2N4waOu1@mail.gmail.com/
Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>