kconfig/satconfig: import PicoSAT 936 source code
In order to support a new make target, "satconfig", we require a SAT
solver. This commit adds PicoSAT, a SAT solver written in C by Armin
Biere and released under a MIT/X11 license.

This commit imports the source code verbatim from:

  <http://fmv.jku.at/picosat/picosat-936.tar.gz>

Please see the PicoSAT website for more information:

  <http://fmv.jku.at/picosat/>

The rest of satconfig will be added in subsequent commits.

Cc: Armin Biere <biere@cs.uni-freiburg.de>
Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>