kconfig/satconf: replace representation of CNF formulae
CNF formulae were represented just like any other boolean expression.
Now we use a list of sparse bitmaps instead, where the list represents
the conjunction and the bitmaps represent the set of literals (in fact
there are two bitmaps for each conjunction; one for the positive
literals and one for the negative literals).

This brings memory usage down to about 320M at peak, and brings the
time for CNF conversion down from about 19 seconds to about 5.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>