kconfig/satconfig: check and warn if configuration changes during writeout
conf_write() and conf_write_autoconf() will perform checks to
(try to) ensure that the configuration is consistent, i.e. it
will try to change the value of options if some constraints
are not satisfied.

This has good and bad sides -- the good is that if it does
change the configuration in any way, then this means that sat-
config did not generate a valid configuration (meaning that
there is an error somewhere in the constraint translation from
kconfig to a sat instance), and this is a quite good way of
debugging such errors.

The bad is that such changes are not always correct, i.e. the
changes themselves may introduce (other) inconsistencies. This
is why we want the SAT solver in the first place.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>