kconfig/satconfig: improve choice block handling
Some things to note: A choice block only requires one of its
choices to be set if at least one of them are visible in the
first place, assuming that the choice block itself is visible.

We also don't handle default values for choice blocks or
choice values, since they are not handled by kconfig either
(attempting to set a default value for a choice value will
trigger a warning saying that this is not supported).

Finally, we currently make choice values depend on their
visibility expression to be true for the symbol value to be
true, instead of just the menu's dependencies.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>