kconfig/satconfig: introduce variables for whether an option was assumed by the user
We need to know whether an option was assumed or not in order to make
decisions about what default to use (defaults should only be applied
where the user didn't assume anything). Since we may want to solve the
same instance with different assumptions, we can't make the encoding
depend on the assumptions.

This also implements defaults for choice blocks.

Signed-off-by: Vegard Nossum <vegard.nossum@gmail.com>