satconfig: include select conditional in sym->selected_expr
When building clauses for whether a symbol has been selected (using the
'select' statement) by a menu entry, we were ignoring the conditional
part of the select statement (i.e. the 'BAR' part of 'select FOO if BAR')
even though we were correctly encoding the conditional part for the
select itself.

Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>