satconfig: remove the 'has at least one prompt if entered by the user' condition
We used to have this in order to avoid generating invalid configs, e.g.:

  config FOO
    default n

could never have FOO=y unless FOO was selected by something -- but we
allowed FOO=y if CONFIG_FOO=y was in the .satconfig.

However, if the fragment had said:

  config BAR
    bool "Bar"

  config FOO
    default y if BAR
    default n

then putting CONFIG_FOO=y in the .satconfig should automatically enable
BAR=y as well.

This is not fully implemented yet, but it's better than giving the "has
at least one prompt if entered by the user" error when trying to set a
value for any option that doesn't have a prompt -- that's one of the
reasons for using satconfig in the first place; the user shouldn't have
to care whether the symbol has a prompt or not (as long as it's possible
to somehow enable it).

Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>