kconfig/satconfig: add basic unit/regression tests
This adds a Python script and some regression tests to aid the
satconfig development and our understanding of the Kconfig language.

Run it like this:

  python scripts/kconfig/satconf-tests.py

Currently looks like this:

  Kconfig.ASN1
   - 0. success
   - 1. success

  Kconfig.CRYPTO_ALGAPI2
   - 0. success

  Kconfig.HAS_IOMEM
   - 0. success
   - 1. success
   - 2. success

  Kconfig.JBD2
   - 0. success

  Kconfig.PAGE_POISONING
   - 0. success

  Kconfig.PREEMPT
   - 0. success
   - 1. success

  Kconfig.SHMEM
   - 0. success
   - 1. success

  Kconfig.WIZNET
   - 0. success

Having this set of example Kconfig files has really helped both my
understanding of how Kconfig works and been invaluable as a tool during
development. It can also serve as a regression suite for future satconfig
(or Kconfig) changes.

Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>