satconfig: fix conditional selects
In the case of conditional selects (like in 'DEBUG_PAGEALLOC select
PAGE_POISONING if !ARCH_SUPPORTS_DEBUG_PAGEALLOC'), we really want
expr_to_bool_expr() to return (x, false). Otherwise, when we negate
the (x, true) value, we get (true, true) when it really should be
(not x, false).

The bug caused us to select PAGE_POISONING when we shouldn't.

Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>