[tmp] kconfig/satconfig: add new "make satmenuconfig" target
This commit adds a proof-of-concept menuconfig-like editor for .satconfig.

The difference from regular menuconfig is that all options are visible at
all times -- no more hunting for dependencies that will make your option
visible.

Of course, you can still run into conflicts (impossible configurations),
but a list of conflicting constraints is presented so that you can easily
find and change the options involved.

Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>